fix: Fix bug in loading .nd2 files with NETTimeLoop
Previously, there was an issue with loading .nd2 files using NETTimeLoop when the field of periods was not included in the parameters. This caused errors and prevented the files from being loaded correctly.

This has been fixed by updating the code to check for the presence of the field of periods and if not, extract the period from other field as necessary.